refactor: extract core extension workspace fact builders
